feat: design tweaks timesheets overview
