+++ assets/scss/_variables.scss
@@ -2,12 +2,17 @@
   --color-primary: #201e33;
   --color-primary-text: #ffffff;
   --color-secondary: #e6e6e6;
-  --color-tertiary: #0cc;
+  --color-tertiary: #62C3DE;
+  --color-tertiary-darker: #21849f;
   --body-color: #212529;
-  --color-success: #28a745;
-  --color-alert: #ff5900;
-  --color-danger: #dc3545;
-  --color-medium-gray: #c1c1c1;
+  --color-success: #A6D9A6;
+  --color-success-darker: #4db34d;
+  --color-alert: #FFB37E;
+  --color-alert-darker: #fe6800;
+  --color-danger: #E8758C;
+  --color-danger-darker: #be203f;
+  --color-medium-gray: #EAEAEA;
+  --color-medium-gray-darker: #aaaaaa;
 }
 
 $theme-colors: (
